requirements.txt
antlr4-python3-runtime==4.13.2
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cssselect2==0.9.0
fonttools==4.62.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
simplejson==3.20.2
stix2==3.0.2
stix2-patterns==2.1.2
tinycss2==1.5.1
tinyhtml5==2.1.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
